chore: Adjust UWP and WinUI specific portions of Window class
