+++ src/Uno.UI/UI/Xaml/Window.Uwp.cs
@@ -1,0 +1,21 @@
+﻿#if !HAS_UNO_WINUI
+#nullable enable
+
+using Windows.UI.ViewManagement;
+
+namespace Windows.UI.Xaml;
+
+partial class Window
+{
+	private bool _windowCreatedRaised;
+
+	partial void RaiseCreated()
+	{
+		if (Application.Current is not null && !_windowCreatedRaised)
+		{
+			_windowCreatedRaised = true;
+			Application.Current.RaiseWindowCreated(this);
+		}
+	}
+}
+#endif
