+++ src/Uno.UI/UI/Xaml/Window.WinAppSdk.cs
@@ -1,5 +1,5 @@
-﻿#nullable enable
-#if HAS_UNO_WINUI
+﻿#if HAS_UNO_WINUI
+#nullable enable
 
 using Windows.UI.ViewManagement;
 
@@ -7,6 +7,12 @@ namespace Windows.UI.Xaml;
 
 public sealed partial class Window
 {
+	public Window() : this(false)
+	{
+	}
+
+	public global::Microsoft.UI.Dispatching.DispatcherQueue DispatcherQueue { get; } = global::Microsoft.UI.Dispatching.DispatcherQueue.GetForCurrentThread();
+
 #if !__IOS__ // This can be added when iOS uses SceneDelegate #8341.
 	/// <summary>
 	/// Gets or sets a string used for the window title.
