+++ src/Uno.UI/UI/Xaml/Window.cs
@@ -29,7 +29,6 @@ namespace Windows.UI.Xaml
 
 		private UIElement _content;
 		private RootVisual _rootVisual;
-		private bool _windowCreatedRaised;
 
 		private CoreWindowActivationState? _lastActivationState;
 		private Brush _background;
@@ -37,16 +36,6 @@ namespace Windows.UI.Xaml
 
 		private List<WeakEventHelper.GenericEventHandler> _sizeChangedHandlers = new List<WeakEventHelper.GenericEventHandler>();
 		private List<WeakEventHelper.GenericEventHandler> _backgroundChangedHandlers;
-
-#if HAS_UNO_WINUI
-		public global::Microsoft.UI.Dispatching.DispatcherQueue DispatcherQueue { get; } = global::Microsoft.UI.Dispatching.DispatcherQueue.GetForCurrentThread();
-#endif
-
-#if HAS_UNO_WINUI
-		public Window() : this(false)
-		{
-		}
-#endif
 
 		internal Window(bool internalUse)
 		{
@@ -97,6 +86,8 @@ namespace Windows.UI.Xaml
 
 			Background = SolidColorBrushHelper.White;
 		}
+
+		partial void RaiseCreated();
 
 		public UIElement Content
 		{
@@ -250,17 +241,6 @@ namespace Windows.UI.Xaml
 				(h, s, e) =>
 					(h as Windows.UI.Xaml.WindowSizeChangedEventHandler)?.Invoke(s, (WindowSizeChangedEventArgs)e)
 			);
-		}
-
-		internal void RaiseCreated()
-		{
-#if !HAS_UNO_WINUI
-			if (Application.Current is not null && !_windowCreatedRaised)
-			{
-				_windowCreatedRaised = true;
-				Application.Current.RaiseWindowCreated(this);
-			}
-#endif
 		}
 
 		internal void OnActivated(CoreWindowActivationState state)
